fix: add types/ namespace for builder, prevent tsc compile extra code
