+++ packages/builder/builder-webpack-provider/package.json
@@ -46,6 +46,9 @@
     },
     "./package.json": {
       "default": "./package.json"
+    },
+    "./types": {
+      "default": "./dist/types/index.js"
     }
   },
   "typesVersions": {
@@ -58,6 +61,9 @@
       ],
       "stub": [
         "./dist/stub/index.d.ts"
+      ],
+      "types": [
+        "./dist/types/index.d.ts"
       ],
       "plugins/*": [
         "./dist/plugins/*.d.ts"
